requirements.txt
Flask
python-docx
docxtpl
docx2pdf
mammoth
